+++ cogs/mod.py
@@ -1,10 +1,11 @@
 import asyncio
 import logging
-from typing import Dict, List
+from typing import Union
 
 import discord
 from discord.ext import commands
 
+from cogs.utils import converter
 from nabbot import NabBot
 from .utils import checks, config, safe_delete_message
 from .utils.context import NabCtx
@@ -15,16 +16,40 @@ from .utils.pages import Pages
 log = logging.getLogger("nabbot")
 
 
+class LazyEntry:
+    __slots__ = ('entity_id', 'guild', '_cache')
+
+    def __init__(self, guild, entity_id):
+        self.entity_id = entity_id
+        self.guild = guild
+        self._cache = None
+
+    def __str__(self):
+        if self._cache:
+            return self._cache
+
+        e = self.entity_id
+        g = self.guild
+        resolved = g.get_channel(e) or g.get_member(e)
+        if resolved is None:
+            self._cache = f'<Not Found: {e}>'
+        else:
+            self._cache = resolved.mention
+        return self._cache
+
+
 class Mod:
-    """Commands server moderators."""
+    """Moderating related commands."""
     def __init__(self, bot: NabBot):
         self.bot = bot
-        self.ignored = {}
-        self.bot.loop.run_until_complete(self.reload_ignored())
 
-    def __global_check(self, ctx: NabCtx):
-        return ctx.is_private or ctx.channel.id not in self.ignored.get(ctx.guild.id, []) or checks.is_owner(ctx) \
-               or checks.check_guild_permissions(ctx, {'manage_channels': True})
+    async def __global_check_once(self, ctx: NabCtx):
+        if ctx.guild is None:
+            return True
+        if await checks.is_owner(ctx):
+            return True
+
+        return await self.is_ignored(ctx.pool, ctx)
 
     # region Commands
     @commands.guild_only()
@@ -56,40 +81,50 @@ class Mod:
 
         await ctx.success(f"Deleted {count:,} messages.", delete_after=20)
 
-    @commands.guild_only()
-    @checks.channel_mod_only()
+    @checks.server_mod_only()
     @commands.group(invoke_without_command=True, case_insensitive=True)
-    async def ignore(self, ctx: NabCtx, *, channel: discord.TextChannel = None):
-        """Makes the bot ignore a channel.
+    async def ignore(self, ctx: NabCtx, *entries: converter.ChannelOrMember):
+        """Makes the bot ignore a channel or user.
 
         Ignored channels don't process commands. However, the bot may still announce deaths and level ups if needed.
+        Ignored users will never get a reply from NabBot.
 
-        If the command   is used with no parameters, it ignores the current channel.
+        If the command is used with no parameters, it ignores the current channel.
 
-        Note that server administrators can bypass this."""
-        if channel is None:
-            channel = ctx.channel
+        Ignore restrictions can be bypassed with the `Manage Guild` permission"""
+        if len(entries) == 0:
+            entries = [ctx.channel]
+        if len(entries) == 1:
+            entry: Union[discord.Member, discord.TextChannel] = entries[0]
+            query = "INSERT INTO ignored_entry(server_id, entry_id) VALUES($1, $2) ON CONFLICT DO NOTHING RETURNING 1"
+            ret = await ctx.pool.fetchval(query, ctx.guild.id, entry.id)
+            rep = entry.mention if isinstance(entry, discord.TextChannel) else entry.display_name
+            if ret:
+                return await ctx.success(f"I'm now ignoring **{rep}**")
+            return await ctx.error(f"{rep} is already ignored.")
 
-        if channel.id in self.ignored.get(ctx.guild.id, []):
-            await ctx.send(f"{channel.mention} is already ignored.")
-            return
+        inserted = await self.bulk_ignore(ctx, entries)
+        if inserted:
+            if inserted != len(entries):
+                await ctx.success(f"{inserted} entries are now ignored. The rest was already ignored.")
+            else:
+                await ctx.success(f"All {inserted} entries are now ignored.")
+        else:
+            await ctx.error("No entries were ignored. They were all already ignored.")
 
-        await ctx.pool.execute("INSERT INTO channel_ignored(server_id, channel_id) VALUES($1, $2)",
-                               ctx.guild.id, channel.id)
-        await ctx.send(f"{channel.mention} is now ignored.")
-        await self.reload_ignored()
-
-    @commands.guild_only()
-    @checks.channel_mod_only()
+    @checks.server_mod_only()
     @ignore.command(name="list")
     async def ignore_list(self, ctx: NabCtx):
-        """Shows a list of ignored channels."""
-        entries = [ctx.guild.get_channel(c).name for c in self.ignored.get(ctx.guild.id, []) if ctx.guild.get_channel(c) is not None]
+        """Shows a list of ignored channels and users."""
+        query = "SELECT entry_id FROM ignored_entry WHERE server_id = $1"
+        rows = await ctx.pool.fetch(query, ctx.guild.id)
+
+        entries = [LazyEntry(ctx.guild, e[0]) for e in rows]
         if not entries:
-            await ctx.send("There are no ignored channels in this server.")
+            await ctx.send("There are no ignored entries in this server.")
             return
         pages = Pages(ctx, entries=entries)
-        pages.embed.title = "Ignored channels"
+        pages.embed.title = "Ignored entries"
         try:
             await pages.paginate()
         except CannotPaginate as e:
@@ -168,24 +203,27 @@ class Mod:
     @commands.guild_only()
     @checks.channel_mod_only()
     @commands.command()
-    async def unignore(self, ctx: NabCtx, *, channel: discord.TextChannel = None):
-        """Unignores a channel.
+    async def unignore(self, ctx: NabCtx, *entries: converter.ChannelOrMember):
+        """Unignores a channel or user..
 
-        If no channel is provided, the current channel will be unignored.
+        If no parameter is provided, the current channel will be unignored.
 
         Ignored channels don't process commands. However, the bot may still announce deaths and level ups if needed.
 
         If the command is used with no parameters, it unignores the current channel."""
-        if channel is None:
-            channel = ctx.channel
-
-        if channel.id not in self.ignored.get(ctx.guild.id, []):
-            await ctx.send(f"{channel.mention} is not ignored.")
-            return
-
-        await ctx.pool.execute("DELETE FROM channel_ignored WHERE channel_id = $1", channel.id)
-        await ctx.send(f"{channel.mention} is not ignored anymore.")
-        await self.reload_ignored()
+        if len(entries) == 0:
+            query = "DELETE FROM ignored_entry WHERE server_id=$1 AND entry_id=$2 RETURNING true"
+            res = await ctx.pool.fetchval(query, ctx.guild.id, ctx.channel.id)
+            if res:
+                return await ctx.success(f"{ctx.channel.mention} is no longer ignored.")
+            return await ctx.error(f"{ctx.channel.mention} wasn't ignored.")
+        query = "DELETE FROM ignored_entry WHERE server_id=$1 AND entry_id = ANY($2::bigint[]) RETURNING entry_id"
+        # noinspection PyUnresolvedReferences
+        entries = [c.id for c in entries]
+        rows = await ctx.pool.fetch(query, ctx.guild.id, entries)
+        if rows:
+            return await ctx.success(f"{len(rows)} are now unignored.")
+        await ctx.error("No entries were unignored.")
 
     @checks.channel_mod_only()
     @checks.tracking_world_only()
@@ -218,21 +256,23 @@ class Mod:
             await ctx.send(e)
     # endregion
 
-    async def reload_ignored(self):
-        """Refresh the world list from the database
+    async def is_ignored(self, conn, ctx: NabCtx):
+        query = "SELECT True FROM ingored_entry WHERE guild_id=$1 AND (entry_id=$2 OR entry_id=$3);"
+        return await conn.fetchrow(query, ctx.guild.id, ctx.channel.id, ctx.author.id)
 
-        This is used to avoid reading the database every time the world list is needed.
-        A global variable holding the world list is loaded on startup and refreshed only when worlds are modified"""
-        ignored_dict_temp: Dict[int, List[int]] = {}
-        result = await self.bot.pool.fetch("SELECT server_id, channel_id FROM channel_ignored")
-        if len(result) > 0:
-            for row in result:
-                if not ignored_dict_temp.get(row["server_id"]):
-                    ignored_dict_temp[row["server_id"]] = []
-                ignored_dict_temp[row["server_id"]].append(row["channel_id"])
+    async def bulk_ignore(self, ctx: NabCtx, entries):
+        async with ctx.pool.acquire() as conn:
+            async with conn.transaction():
+                query = "SELECT entry_id FROM ignored_entry WHERE server_id=$1;"
+                records = await conn.fetch(query, ctx.guild.id)
 
-        self.ignored.clear()
-        self.ignored.update(ignored_dict_temp)
+                # Removing duplicates
+                current_entries = {r[0] for r in records}
+                records = [(ctx.guild.id, e.id) for e in entries if e.id not in current_entries]
+
+                # do a bulk COPY
+                await conn.copy_records_to_table('ignored_entry', columns=('server_id', 'entry_id'), records=records)
+                return len(records)
 
 
 def setup(bot):